requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.8.1
async-generator==1.10
attrs==23.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
Brlapi==0.8.3
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.4
djangorestframework==3.15.2
duplicity==0.8.21
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fasteners==0.14.1
future==0.18.2
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.10
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.62.0
grpcio==1.68.1
grpcio-status==1.62.3
gspread==6.0.2
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.2.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
proto-plus==1.25.0
protobuf==4.25.3
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.3
pydantic_core==2.27.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.2
StrEnum==0.4.15
systemd-python==234
tabulate==0.9.0
telegram-export==1.8.2
Telethon==1.4.3
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
